requirements.txt
python-dotenv
asana
google-cloud-bigquery
google-api-python-client
google-auth-httplib2
pandas